chore(INFRA-831): move application to a new namespace (python3_saml) to avoid conflicting with onelogin
BREAKING CHANGE: imports will need to be changed from `from onelogin.<path> import ...` to `from python3_saml.<path> import ...`
