+++ demo-tornado/views.py
@@ -4,8 +4,8 @@ import Settings
 import tornado.httpserver
 import tornado.httputil
 
-from onelogin.saml2.auth import OneLogin_Saml2_Auth
-from onelogin.saml2.utils import OneLogin_Saml2_Utils
+from python3_saml.saml2.auth import OneLogin_Saml2_Auth
+from python3_saml.saml2.utils import OneLogin_Saml2_Utils
 
 # Global session info
 session = {}
