+++ demo_pyramid/demo_pyramid/views.py
@@ -7,8 +7,8 @@ from pyramid.httpexceptions import (
 )
 from pyramid.view import view_config
 
-from onelogin.saml2.auth import OneLogin_Saml2_Auth
-from onelogin.saml2.utils import OneLogin_Saml2_Utils
+from python3_saml.saml2.auth import OneLogin_Saml2_Auth
+from python3_saml.saml2.utils import OneLogin_Saml2_Utils
 
 SAML_PATH = os.path.join(os.path.dirname(__file__), "saml")
