+++ src/python3_saml/saml2/compat.py
