+++ src/python3_saml/saml2/idp_metadata_parser.py
@@ -10,8 +10,8 @@ from urllib.request import Request, urlopen
 
 import ssl
 
-from onelogin.saml2.constants import OneLogin_Saml2_Constants
-from onelogin.saml2.xml_utils import OneLogin_Saml2_XML
+from python3_saml.saml2.constants import OneLogin_Saml2_Constants
+from python3_saml.saml2.xml_utils import OneLogin_Saml2_XML
 
 
 class OneLogin_Saml2_IdPMetadataParser(object):
