+++ src/python3_saml/saml2/logout_request.py
@@ -7,11 +7,11 @@ Logout Request class of SAML Python Toolkit.
 
 """
 
-from onelogin.saml2 import compat
-from onelogin.saml2.constants import OneLogin_Saml2_Constants
-from onelogin.saml2.utils import OneLogin_Saml2_Utils, OneLogin_Saml2_Error, OneLogin_Saml2_ValidationError
-from onelogin.saml2.xml_templates import OneLogin_Saml2_Templates
-from onelogin.saml2.xml_utils import OneLogin_Saml2_XML
+from python3_saml.saml2 import compat
+from python3_saml.saml2.constants import OneLogin_Saml2_Constants
+from python3_saml.saml2.utils import OneLogin_Saml2_Utils, OneLogin_Saml2_Error, OneLogin_Saml2_ValidationError
+from python3_saml.saml2.xml_templates import OneLogin_Saml2_Templates
+from python3_saml.saml2.xml_utils import OneLogin_Saml2_XML
 
 
 class OneLogin_Saml2_Logout_Request(object):
