+++ src/python3_saml/saml2/metadata.py
@@ -10,11 +10,11 @@ Metadata class of SAML Python Toolkit.
 from time import gmtime, strftime, time
 from datetime import datetime
 
-from onelogin.saml2 import compat
-from onelogin.saml2.constants import OneLogin_Saml2_Constants
-from onelogin.saml2.utils import OneLogin_Saml2_Utils
-from onelogin.saml2.xml_templates import OneLogin_Saml2_Templates
-from onelogin.saml2.xml_utils import OneLogin_Saml2_XML
+from python3_saml.saml2 import compat
+from python3_saml.saml2.constants import OneLogin_Saml2_Constants
+from python3_saml.saml2.utils import OneLogin_Saml2_Utils
+from python3_saml.saml2.xml_templates import OneLogin_Saml2_Templates
+from python3_saml.saml2.xml_utils import OneLogin_Saml2_XML
 
 try:
     basestring
