+++ src/python3_saml/saml2/schemas/saml-schema-assertion-2.0.xsd
