+++ src/python3_saml/saml2/schemas/sstc-metadata-attr.xsd
