+++ src/python3_saml/saml2/schemas/sstc-saml-metadata-algsupport-v1.0.xsd
