+++ src/python3_saml/saml2/schemas/xmldsig-core-schema.xsd
