+++ tests/src/Python3_Saml/saml2_tests/logout_request_test.py
@@ -6,10 +6,10 @@ from os.path import dirname, join, exists
 import unittest
 from xml.dom.minidom import parseString
 
-from onelogin.saml2 import compat
-from onelogin.saml2.logout_request import OneLogin_Saml2_Logout_Request
-from onelogin.saml2.settings import OneLogin_Saml2_Settings
-from onelogin.saml2.utils import OneLogin_Saml2_Utils
+from python3_saml.saml2 import compat
+from python3_saml.saml2.logout_request import OneLogin_Saml2_Logout_Request
+from python3_saml.saml2.settings import OneLogin_Saml2_Settings
+from python3_saml.saml2.utils import OneLogin_Saml2_Utils
 
 try:
     from urllib.parse import urlparse, parse_qs
