+++ tests/src/Python3_Saml/saml2_tests/logout_response_test.py
@@ -6,12 +6,12 @@ from os.path import dirname, join, exists
 import unittest
 from xml.dom.minidom import parseString
 
-from onelogin.saml2 import compat
-from onelogin.saml2.constants import OneLogin_Saml2_Constants
-from onelogin.saml2.logout_response import OneLogin_Saml2_Logout_Response
-from onelogin.saml2.settings import OneLogin_Saml2_Settings
-from onelogin.saml2.utils import OneLogin_Saml2_Utils
-from onelogin.saml2.utils import OneLogin_Saml2_XML
+from python3_saml.saml2 import compat
+from python3_saml.saml2.constants import OneLogin_Saml2_Constants
+from python3_saml.saml2.logout_response import OneLogin_Saml2_Logout_Response
+from python3_saml.saml2.settings import OneLogin_Saml2_Settings
+from python3_saml.saml2.utils import OneLogin_Saml2_Utils
+from python3_saml.saml2.utils import OneLogin_Saml2_XML
 
 try:
     from urllib.parse import urlparse, parse_qs
