+++ tests/src/Python3_Saml/saml2_tests/metadata_test.py
@@ -7,12 +7,12 @@ from time import strftime
 from datetime import datetime, timezone
 import unittest
 
-from onelogin.saml2 import compat
-from onelogin.saml2.metadata import OneLogin_Saml2_Metadata
-from onelogin.saml2.settings import OneLogin_Saml2_Settings
-from onelogin.saml2.constants import OneLogin_Saml2_Constants
-from onelogin.saml2.utils import OneLogin_Saml2_Utils
-from onelogin.saml2.xml_utils import OneLogin_Saml2_XML
+from python3_saml.saml2 import compat
+from python3_saml.saml2.metadata import OneLogin_Saml2_Metadata
+from python3_saml.saml2.settings import OneLogin_Saml2_Settings
+from python3_saml.saml2.constants import OneLogin_Saml2_Constants
+from python3_saml.saml2.utils import OneLogin_Saml2_Utils
+from python3_saml.saml2.xml_utils import OneLogin_Saml2_XML
 
 
 class OneLogin_Saml2_Metadata_Test(unittest.TestCase):
